refactor: rename substrate config interface to WsProviderURLGetter
